# lib/api-zod/src/generated/types/columnStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sales Forecasting System API
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnStatsHistogramItem } from './columnStatsHistogramItem';

export interface ColumnStats {
  column: string;
  /** @nullable */
  mean: number | null;
  /** @nullable */
  median: number | null;
  /** @nullable */
  std: number | null;
  /** @nullable */
  min: number | null;
  /** @nullable */
  max: number | null;
  /** @nullable */
  q25: number | null;
  /** @nullable */
  q75: number | null;
  histogram: ColumnStatsHistogramItem[];
}