# lib/api-zod/src/generated/types/compareResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sales Forecasting System API
 * OpenAPI spec version: 0.1.0
 */
import type { CompareResultModelsItem } from './compareResultModelsItem';

export interface CompareResult {
  models: CompareResultModelsItem[];
}