# lib/api-zod/src/generated/types/correlationMatrix.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sales Forecasting System API
 * OpenAPI spec version: 0.1.0
 */

export interface CorrelationMatrix {
  columns: string[];
  values: number[][];
}